move more classes out of org.apache.pekko package
