+++ examples/app/src/main/resources/application.conf
@@ -11,8 +11,8 @@ pekko {
     provider = "org.apache.pekko.cluster.ClusterActorRefProvider"
     allow-java-serialization = true // for in-memory unit tests only!
     default-dispatcher {
-      type = "org.apache.pekko.sensors.dispatch.InstrumentedDispatcherConfigurator"
-      executor = "org.apache.pekko.sensors.dispatch.InstrumentedExecutor"
+      type = "nl.pragmasoft.pekko.sensors.dispatch.InstrumentedDispatcherConfigurator"
+      executor = "nl.pragmasoft.pekko.sensors.dispatch.InstrumentedExecutor"
 
       instrumented-executor {
         delegate = "cats.effect.unsafe.CatsEffectPool"
@@ -24,8 +24,8 @@ pekko {
     }
 
     default-blocking-io-dispatcher {
-      type = "org.apache.pekko.sensors.dispatch.InstrumentedDispatcherConfigurator"
-      executor = "org.apache.pekko.sensors.dispatch.InstrumentedExecutor"
+      type = "nl.pragmasoft.pekko.sensors.dispatch.InstrumentedDispatcherConfigurator"
+      executor = "nl.pragmasoft.pekko.sensors.dispatch.InstrumentedExecutor"
 
       instrumented-executor {
         delegate = "thread-pool-executor"
