+++ sensors/src/main/scala/nl/pragmasoft/pekko/sensors/metered/MeteredDispatcherWrapper.scala
@@ -1,10 +1,10 @@
 package nl.pragmasoft.pekko.sensors.metered
 
-import org.apache.pekko.sensors.dispatch.DispatcherInstrumentationWrapper.{InstrumentedRun, Run}
 import com.typesafe.config.Config
 import nl.pragmasoft.pekko.sensors.RunnableWatcher
+import nl.pragmasoft.pekko.sensors.dispatch.DispatcherInstrumentationWrapper.{InstrumentedRun, Run}
+import nl.pragmasoft.pekko.sensors.dispatch.RunnableWrapper
 import org.apache.pekko.sensors.DispatcherMetrics
-import org.apache.pekko.sensors.dispatch.RunnableWrapper
 
 import java.util.concurrent.atomic.LongAdder
 import scala.concurrent.duration.Duration
